renamed dataconnection folder to config
